backend/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
pydantic
pyjwt
bcrypt
chromadb
pytest
httpx
